Revert "data 관련 파일 디렉토리 정리 및 RDB 데이터 연동"
